# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"Workflow Management System","version":"0.1.0"},"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/my":{"get":{"tags":["Tasks – Designer"],"summary":"My tasks (Designer)","operationId":"my_tasks_tasks_my_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Response My Tasks Tasks My Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/tasks/pending-approval":{"get":{"tags":["Tasks – Approval"],"summary":"Pending approval tasks","operationId":"pending_approval_tasks_tasks_pending_approval_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Response Pending Approval Tasks Tasks Pending Approval Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/tasks/":{"get":{"tags":["Tasks – Core"],"summary":"List Tasks","operationId":"list_tasks_tasks__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TaskResponse"},"type":"array","title":"Response List Tasks Tasks  Get"}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["Tasks – Core"],"summary":"Create Task","operationId":"create_task_tasks__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/tasks/{task_id}":{"get":{"tags":["Tasks – Core"],"summary":"Get Task","operationId":"get_task_tasks__task_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/assign":{"post":{"tags":["Tasks – Assignment (Admin)"],"summary":"Assign Designer","operationId":"assign_designer_tasks__task_id__assign_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignDesignerRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/start":{"post":{"tags":["Tasks – Designer"],"summary":"Start Task","operationId":"start_task_tasks__task_id__start_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/complete":{"post":{"tags":["Tasks – Designer"],"summary":"Complete Task","operationId":"complete_task_tasks__task_id__complete_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/approve":{"post":{"tags":["Tasks – Approval"],"summary":"Approve Task","operationId":"approve_task_tasks__task_id__approve_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/request-changes":{"post":{"tags":["Tasks – Approval"],"summary":"Request Changes","operationId":"request_changes_tasks__task_id__request_changes_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestChangesRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/history/":{"get":{"tags":["Tasks – History"],"summary":"Get Task History","operationId":"get_task_history_tasks__task_id__history__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskHistoryResponse"},"title":"Response Get Task History Tasks  Task Id  History  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/{task_id}/comments/":{"post":{"tags":["Tasks – Comments"],"summary":"Add comment to task","operationId":"add_comment_tasks__task_id__comments__post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCommentCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskCommentResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Tasks – Comments"],"summary":"List task comments","operationId":"list_comments_tasks__task_id__comments__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskCommentResponse"},"title":"Response List Comments Tasks  Task Id  Comments  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/notifications/":{"get":{"tags":["Notifications"],"summary":"My Notifications","operationId":"my_notifications_notifications__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/NotificationResponse"},"type":"array","title":"Response My Notifications Notifications  Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/notifications/{notification_id}/read":{"post":{"tags":["Notifications"],"summary":"Mark Notification Read","operationId":"mark_notification_read_notifications__notification_id__read_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"notification_id","in":"path","required":true,"schema":{"type":"string","title":"Notification Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/tasks/import/preview":{"post":{"tags":["Tasks – Import"],"summary":"Upload CSV/Excel and preview tasks","operationId":"preview_task_import_tasks_import_preview_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_preview_task_import_tasks_import_preview_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskImportPreviewResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/tasks/import/validate":{"post":{"tags":["Tasks – Import"],"summary":"Validate imported task rows","operationId":"validate_task_import_tasks_import_validate_post","requestBody":{"content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Rows"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskImportValidationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/tasks/import/commit":{"post":{"tags":["Tasks – Import"],"summary":"Commit validated tasks to database","operationId":"commit_task_import_tasks_import_commit_post","requestBody":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TaskImportCommitRow"},"type":"array","title":"Rows"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskImportCommitResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/analytics/performance/designers":{"get":{"tags":["Analytics – Performance"],"summary":"Performance of all designers (Admin)","operationId":"performance_all_designers_analytics_performance_designers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DesignerPerformance"},"type":"array","title":"Response Performance All Designers Analytics Performance Designers Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/analytics/performance/me":{"get":{"tags":["Analytics – Performance"],"summary":"My performance (Designer)","operationId":"my_performance_analytics_performance_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MyPerformance"}}}}},"security":[{"HTTPBearer":[]}]}},"/analytics/bottlenecks/overdue":{"get":{"tags":["Analytics – Bottlenecks"],"summary":"Overdue tasks (Admin)","operationId":"overdue_tasks_analytics_bottlenecks_overdue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/TaskRisk"},"type":"array","title":"Response Overdue Tasks Analytics Bottlenecks Overdue Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/analytics/bottlenecks/at-risk":{"get":{"tags":["Analytics – Bottlenecks"],"summary":"Tasks nearing deadline","operationId":"tasks_at_risk_analytics_bottlenecks_at_risk_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":3,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskRisk"},"title":"Response Tasks At Risk Analytics Bottlenecks At Risk Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/analytics/bottlenecks/designer-load":{"get":{"tags":["Analytics – Bottlenecks"],"summary":"Overloaded designers","operationId":"designer_load_analytics_bottlenecks_designer_load_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"threshold","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Threshold"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DesignerLoad"},"title":"Response Designer Load Analytics Bottlenecks Designer Load Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/analytics/bottlenecks/stuck":{"get":{"tags":["Analytics – Bottlenecks"],"summary":"Tasks stuck in working state","operationId":"stuck_tasks_analytics_bottlenecks_stuck_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"days","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Days"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TaskRisk"},"title":"Response Stuck Tasks Analytics Bottlenecks Stuck Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Health","operationId":"health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ApprovalStatus":{"type":"string","enum":["Pending","Approved","ChangesRequired","Rejected","OnHold"],"title":"ApprovalStatus"},"AssignDesignerRequest":{"properties":{"designer_id":{"type":"string","title":"Designer Id"}},"type":"object","required":["designer_id"],"title":"AssignDesignerRequest"},"Body_preview_task_import_tasks_import_preview_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_preview_task_import_tasks_import_preview_post"},"DesignStatus":{"type":"string","enum":["Pending","Working","OnHold","Completed","Discarded","NotCompleted"],"title":"DesignStatus"},"DesignerLoad":{"properties":{"designer_id":{"type":"string","title":"Designer Id"},"active_tasks":{"type":"integer","title":"Active Tasks"}},"type":"object","required":["designer_id","active_tasks"],"title":"DesignerLoad"},"DesignerPerformance":{"properties":{"designer_id":{"type":"string","title":"Designer Id"},"completed":{"type":"integer","title":"Completed"},"pending":{"type":"integer","title":"Pending"},"total":{"type":"integer","title":"Total"},"completion_rate":{"type":"number","title":"Completion Rate"}},"type":"object","required":["designer_id","completed","pending","total","completion_rate"],"title":"DesignerPerformance"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginRequest"},"MyPerformance":{"properties":{"completed":{"type":"integer","title":"Completed"},"pending":{"type":"integer","title":"Pending"},"total":{"type":"integer","title":"Total"},"completion_rate":{"type":"number","title":"Completion Rate"}},"type":"object","required":["completed","pending","total","completion_rate"],"title":"MyPerformance"},"NotificationResponse":{"properties":{"id":{"type":"string","title":"Id"},"user_id":{"type":"string","title":"User Id"},"type":{"type":"string","title":"Type"},"message":{"type":"string","title":"Message"},"task_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task Id"},"is_read":{"type":"boolean","title":"Is Read"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","user_id","type","message","is_read","created_at"],"title":"NotificationResponse"},"PostingStatus":{"type":"string","enum":["Draft","Scheduled","Posted","Failed"],"title":"PostingStatus"},"RequestChangesRequest":{"properties":{"comment":{"type":"string","title":"Comment"}},"type":"object","required":["comment"],"title":"RequestChangesRequest"},"TaskCommentCreate":{"properties":{"content":{"type":"string","minLength":1,"title":"Content","description":"Comment text"}},"type":"object","required":["content"],"title":"TaskCommentCreate"},"TaskCommentResponse":{"properties":{"id":{"type":"string","title":"Id"},"task_id":{"type":"string","title":"Task Id"},"author_id":{"type":"string","title":"Author Id"},"author_role":{"type":"string","title":"Author Role"},"content":{"type":"string","title":"Content"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","task_id","author_id","author_role","content","created_at"],"title":"TaskCommentResponse"},"TaskCreate":{"properties":{"content_type":{"type":"string","title":"Content Type","description":"Content Type from Excel"},"size":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Size"},"title":{"type":"string","title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"instructions":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Instructions"},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]}},"type":"object","required":["content_type","title"],"title":"TaskCreate"},"TaskHistoryResponse":{"properties":{"id":{"type":"string","title":"Id"},"task_id":{"type":"string","title":"Task Id"},"action":{"type":"string","title":"Action"},"performed_by":{"type":"string","title":"Performed By"},"role":{"type":"string","title":"Role"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","task_id","action","performed_by","role","created_at"],"title":"TaskHistoryResponse"},"TaskImportCommitResponse":{"properties":{"total_rows":{"type":"integer","title":"Total Rows"},"created":{"type":"integer","title":"Created"},"failed":{"type":"integer","title":"Failed"},"failed_rows":{"items":{"type":"integer"},"type":"array","title":"Failed Rows"}},"type":"object","required":["total_rows","created","failed","failed_rows"],"title":"TaskImportCommitResponse"},"TaskImportCommitRow":{"properties":{"row_number":{"type":"integer","title":"Row Number"},"data":{"additionalProperties":true,"type":"object","title":"Data"}},"type":"object","required":["row_number","data"],"title":"TaskImportCommitRow"},"TaskImportPreviewResponse":{"properties":{"total_rows":{"type":"integer","title":"Total Rows"},"rows":{"items":{"$ref":"#/components/schemas/TaskImportRow"},"type":"array","title":"Rows"}},"type":"object","required":["total_rows","rows"],"title":"TaskImportPreviewResponse"},"TaskImportRow":{"properties":{"row_number":{"type":"integer","title":"Row Number"},"data":{"additionalProperties":true,"type":"object","title":"Data"}},"type":"object","required":["row_number","data"],"title":"TaskImportRow"},"TaskImportValidationError":{"properties":{"row_number":{"type":"integer","title":"Row Number"},"errors":{"items":{"type":"string"},"type":"array","title":"Errors"}},"type":"object","required":["row_number","errors"],"title":"TaskImportValidationError"},"TaskImportValidationResponse":{"properties":{"total_rows":{"type":"integer","title":"Total Rows"},"valid_rows":{"type":"integer","title":"Valid Rows"},"invalid_rows":{"type":"integer","title":"Invalid Rows"},"errors":{"items":{"$ref":"#/components/schemas/TaskImportValidationError"},"type":"array","title":"Errors"}},"type":"object","required":["total_rows","valid_rows","invalid_rows","errors"],"title":"TaskImportValidationResponse"},"TaskResponse":{"properties":{"content_type":{"type":"string","title":"Content Type","description":"Content Type from Excel"},"size":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Size"},"title":{"type":"string","title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"instructions":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Instructions"},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"id":{"type":"string","title":"Id"},"assigned_by_id":{"type":"string","title":"Assigned By Id"},"designer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Designer Id"},"design_status":{"$ref":"#/components/schemas/DesignStatus"},"approval_status":{"$ref":"#/components/schemas/ApprovalStatus"},"posting_status":{"$ref":"#/components/schemas/PostingStatus"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["content_type","title","id","assigned_by_id","designer_id","design_status","approval_status","posting_status","created_at","updated_at"],"title":"TaskResponse"},"TaskRisk":{"properties":{"task_id":{"type":"string","title":"Task Id"},"title":{"type":"string","title":"Title"},"designer_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Designer Id"},"deadline":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deadline"},"days_remaining":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Days Remaining"}},"type":"object","required":["task_id","title","designer_id","deadline","days_remaining"],"title":"TaskRisk"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"UserCreate":{"properties":{"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role"}},"type":"object","required":["full_name","email","password","role"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}